refactoring: Moved fixtures to conftest.py to be able to reuse them in other test_files.
